line-editorの全機能を実装。でも多分バグだらけ。

+++ Chap11/line-editor/command-parser.h
@@ -22,7 +22,7 @@ typedef enum {
 typedef struct {
   bool valid;
   COMMAND_TYPE type;
-  int line;
+  long line;
   char *path;
 } COMMAND;
